+++ geometry-helpers.h
@@ -9,4 +9,4 @@ typedef struct
     float y;
 } Point_t;
 
-int32_t areLinesCrossing(Point_t p1, Point_t p2,  Point_t v1, Point_t v2);
+float areLinesCrossing(Point_t p1, Point_t p2,  Point_t v1, Point_t v2);
